+++ packages/react-core/src/operations/constants/Provider.ts
@@ -3,5 +3,6 @@ export enum Provider {
   Redshift = 'redshift',
   Postgres = 'postgres',
   Snowflake = 'snowflake',
-  Databricks = 'databricks'
+  Databricks = 'databricks',
+  DatabricksRest = 'databricksRest'
 }